add tests

+++ src/repositories/blogs-repository-mongo.ts
@@ -5,11 +5,13 @@ export interface IBlog {
     id: ObjectId;
     name: string;
     youtubeUrl: string;
+    createdAt: string;
 }
 
 export interface IBlogMongo  {
     name: string;
     youtubeUrl: string;
+    createdAt: string;
 }
 
 export const blogsRepository = {
@@ -19,18 +21,23 @@ export const blogsRepository = {
             return {
                 id: item._id,
                 name: item.name,
-                youtubeUrl: item.youtubeUrl
+                youtubeUrl: item.youtubeUrl,
+                createdAt: item.youtubeUrl,
             }
         });
     },
 
     async getSingleBlog(id: string): Promise<IBlog | null> {
+        if (!ObjectId.isValid(id)) {
+            return null;
+        }
         const result = await connectDbBlogs.findOne({_id: new ObjectId(id)});
         if (result) {
             return {
                 id: result._id,
                 name: result.name,
-                youtubeUrl: result.youtubeUrl
+                youtubeUrl: result.youtubeUrl,
+                createdAt: result.createdAt
             };
         } else {
             return null;
@@ -38,13 +45,14 @@ export const blogsRepository = {
     },
 
     async createBlog(name: string, youtubeUrl:string): Promise<IBlog | null> {
-        const result = await connectDbBlogs.insertOne({ name, youtubeUrl});
+        const result = await connectDbBlogs.insertOne({ name, youtubeUrl, createdAt: new Date().toISOString()});
         const blog = await connectDbBlogs.findOne({_id: result.insertedId});
         if (blog) {
             return {
                 id: blog._id,
                 name: blog.name,
-                youtubeUrl: blog.youtubeUrl
+                youtubeUrl: blog.youtubeUrl,
+                createdAt: new Date().toISOString()
             };
         } else {
             return null;
@@ -52,11 +60,17 @@ export const blogsRepository = {
     },
 
     async changeBlog(id: string, name: string, youtubeUrl: string):Promise<boolean> {
+        if (!ObjectId.isValid(id)) {
+            return false;
+        }
         const result = await connectDbBlogs.updateOne({_id: new ObjectId(id)}, {$set: {name, youtubeUrl}})
         return result.acknowledged;
     },
 
     async deleteBlog(id: string):Promise<boolean> {
+        if (!ObjectId.isValid(id)) {
+            return false;
+        }
         const result = await connectDbBlogs.deleteOne({_id: new ObjectId(id)});
         return result.acknowledged;
     },
